Modified thumbnailType support for more flexibility
Given the name of the option, it seems backwards to only support it for PDF thumbnails. This commit changes it's support such that when we specify a `thumbnailType`, this is used for both thumbnails and pdfs.

It also adds a separate option, `mediaThumbnailType` for dealing specifically with non-image media files. This could include PDF files, AVI files etc.