Allow to specify sub folders of thumb size. Closes #32
Uses the `{size}` and `{geometry}` key to figure these two out. Also added a way to configure thumbnail renaming as it might not be useful to some people to have the `{size}` displayed both in the directory path and the filename.

Also renamed two options, `thumbsizes` and `thumbnailPrefixStyle`, in order to consolidate the thumbnail settings