Added folder removal option
As I commented a few weeks ago, an option to kill the folder (if the user desires it) is useful. Helps keep the server clean.

I'm not sure if I've done this in the cleanest way. For example, the normal deletion process runs as normal, I simply trash the folder (and any remaining contents) in afterDelete.

Of course we could completely refactor the whole deletion procedure, but I wanted to make as few changes as possible.