Fixed randomCombined to save model->id in path and added test
Until now the model->id was not part of the saved path in the
randomCombined pathMethod, its value was only being used to calculate
the random path numbers and a test for this method was missing.

Also a small addition on README to explain the format produced by the
two random pathMethods.