Add support for field-value replacements
The pattern `{field-value:(\w+)}` is replaced by value contained in the current entity in the specified field. As an example, if your path has `{field-value:unique_id}` and the entity being saved has a value of `4b3403665fea6` for the field `unique_id`, then `{field-value:unique_id}` will be replaced with `4b3403665fea6`. This replacement can be used multiple times for one or more fields.

If the value is not a string or zero-length, a LogicException will be thrown.