Make files stored in non-static paths deletable.
Reintroduces the usage of the stored path on delete (#390). The
stored path is preferred over the dynamically generated path, as the
latter might be different to the path on disk, hence causing files to
be not deletable.

The stored path is only used in case it's not `null`. This should ensure
backwards compatibility for applications that do not store the path, and
expect the path processor to be used.

refs #409