Update docs.
Made it more clear when and why storing the path in the record is
required in order for files to be deletable.

Removed the form example as there is no need to pass on the path,
neither should users be able to modify that value, nor is the form
required in the context of deleting records/files.

Also changed the form example to use proper arguments, and follow the
more popular usage of not providing the base table name for the
inputs.